Replace Bash script with a Python script.
There were several problems with the Bash script:

The first is that I failed to read one field so many fields were incorrectly matched.

The second is that the `read` Bash builtin always treats the consecutive whitespace characters space and tab as a single delimiter. Working around this would have been difficult and the need to do so inspired me to write a Python script instead.

The third is that the `read` builtin was escaping extended characters. I think this would have been easy to fix but switching to Python essentially fixed it 'for free'.

I also made changes to the function of the new version of the script:

 1. I removed the line that repeated the password.
 2. I removed the 'password policy' fields as they seem to use an internal representation only understandable by pwSafe.
 3. I removed the 'history' field as it also seems to use an internal representation only understandable by pwSafe.
 4. I added a newline between the \"Notes:\" field label and the field contents.

Case 1349